NFS mount documentation, locking issue
I ran into some issues when I ran software that was trying to lock files under the nfs folder. This was on Catalina and looks like this is a common issue with Mac -> Linux nfs. 

After digging into it, for my use case using the `locallocks` NFS option on the Mac client resolved it. The `locallocks` option means that locking is handled by the client. There would be problems if you actually need server side locking.